stop and remove docker container only if it exists

+++ {{cookiecutter.project_slug}}/tests/CI_docker/build_{{cookiecutter.project_slug}}_testsuite_image.sh
@@ -29,8 +29,12 @@ docker build ${context_dir} \
 #       The runner uses a 'config.toml' configuration file at /etc/gitlab-runner within the container which can be
 #       modified through additional parameters of the 'gitlab-runner register' command.
 echo "#### Create gitlab-runner (daemon) container with tag; ${tag}"
-docker stop ${gitlab_runner}
-docker rm ${gitlab_runner}
+if [ "$(docker ps -qa -f name=${gitlab_runner})" ]; then
+    if [ "$(docker ps -q -f name=${gitlab_runner})" ]; then
+        docker stop ${gitlab_runner};
+    fi
+    docker rm ${gitlab_runner};
+fi
 docker run \
     -d \
     --name ${gitlab_runner} \
